Refactor(analysis): separate actual result and result value
