+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/result/AbstractPerformanceResult.java
@@ -1,7 +1,5 @@
 package de.fabiankeller.palladio.analysis.result;
 
-import de.fabiankeller.palladio.analysis.result.type.Result;
-
 import java.util.*;
 import java.util.stream.Collectors;
