+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/result/Result.java
@@ -1,4 +1,6 @@
-package de.fabiankeller.palladio.analysis.result.type;
+package de.fabiankeller.palladio.analysis.result;
+
+import de.fabiankeller.palladio.analysis.result.valueobject.ValueObject;
 
 /**
  * Represents a single performance analysis result, that is specific for a element in the PCM.
@@ -11,4 +13,9 @@ public interface Result<T> {
      * Returns the element this result is attached to.
      */
     T attachedTo();
+
+    /**
+     * The actual value.
+     */
+    ValueObject value();
 }
