+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/result/type/Throughput.java
@@ -1,20 +1,25 @@
 package de.fabiankeller.palladio.analysis.result.type;
 
 import de.fabiankeller.palladio.analysis.result.valueobject.NormalPercentage;
+import de.fabiankeller.palladio.analysis.result.valueobject.ValueObject;
 
 /**
  * Defines the throughput of a constrained resource.
  */
-public class Throughput<T> extends AbstractResult<T> {
-    
+public class Throughput<T> implements ValueObject {
+
     private final NormalPercentage throughput;
 
-    public Throughput(final T attachedTo1, final NormalPercentage throughput) {
-        super(attachedTo1);
+    public Throughput(final NormalPercentage throughput) {
         this.throughput = throughput;
     }
 
     public NormalPercentage getThroughput() {
         return this.throughput;
+    }
+
+    @Override
+    public String toHumanReadable() {
+        return "Throughput: " + this.throughput.toHumanReadable();
     }
 }
